Extract RootCommand class
