+++ lib/mli/cli/root_command.rb
@@ -1,0 +1,20 @@
+require "thor"
+
+module Mli
+  module Cli
+    class RootCommand < Thor
+      desc "version", "Print the version number."
+      def version
+        say Mli::VERSION
+      end
+
+      def self.basename
+        "mli"
+      end
+
+      def self.exit_on_failure?
+        true
+      end
+    end
+  end
+end
